requirements.txt
aniso8601==3.0.0
astroid==1.6.2
click==6.7
flake8==3.5.0
Flask==0.12.2
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.3
pytz==2018.4
PyYAML==3.12
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
uwsgi
